Time: 2 ms (73.07%), Space: 14.3 MB (55.28%) - LeetHub

+++ 1482-how-many-numbers-are-smaller-than-the-current-number/1482-how-many-numbers-are-smaller-than-the-current-number.cpp
@@ -1,15 +1,15 @@
 class Solution {
 public:
     vector<int> smallerNumbersThanCurrent(vector<int>& nums) {
-        unsigned short numsSize = nums.size();
+        int numsSize = nums.size();
         vector<int> answer;
         answer.reserve(numsSize);
 
-        for (unsigned short i = 0; i < numsSize; ++i)
+        for (int i = 0; i < numsSize; ++i)
         {
-            unsigned short count = 0;
+            int count = 0;
 
-            for (unsigned short j = 0; j < numsSize; ++j)
+            for (int j = 0; j < numsSize; ++j)
                 if (nums[i] > nums[j])
                     ++count;
